[WIP] Reverse diff for functional inversions
Work in progress commit. Still errors due to problems with mutable structs. We need to decide to go either for ReverseDiff with PreallocationTools or with Enzyme.

+++ src/setup/config.jl
@@ -53,7 +53,7 @@ Initializes ODINN by configuring PyCall based on a given Python path. It also co
 for a given number of processes. 
 """
 function enable_multiprocessing(params::Sleipnir.Parameters)
-    procs::Int = params.simulation.workers
+    procs = params.simulation.workers
     if procs > 0 && params.simulation.multiprocessing
         if nprocs() < procs
             @eval begin
